Refactor contact create models
