+++ epp-client/src/epp.rs
@@ -5,7 +5,6 @@ pub mod request;
 pub mod response;
 pub mod xml;
 
-pub use request::contact::create::*;
 pub use request::contact::delete::*;
 pub use request::contact::info::*;
 pub use request::contact::update::*;
@@ -17,7 +16,6 @@ pub use request::host::update::*;
 pub use request::message::ack::*;
 pub use request::message::poll::*;
 
-pub use response::contact::create::*;
 pub use response::contact::delete::*;
 pub use response::contact::info::*;
 pub use response::contact::update::*;
