+++ epp-client/src/epp/request/contact.rs
@@ -1,6 +1,5 @@
 //! Types for EPP contact requests
 
-pub mod create;
 pub mod delete;
 pub mod info;
 pub mod update;
