+++ epp-client/src/epp/response/contact.rs
@@ -1,6 +1,5 @@
 //! Types for EPP contact responses
 
-pub mod create;
 pub mod delete;
 pub mod info;
 pub mod update;
