+++ epp-client/src/tests/se.rs
@@ -4,6 +4,7 @@ mod request {
     use super::super::get_xml;
     use super::super::CLTRID;
     use crate::contact::check::EppContactCheck;
+    use crate::contact::create::EppContactCreate;
     use crate::domain::check::EppDomainCheck;
     use crate::domain::create::EppDomainCreate;
     use crate::domain::delete::EppDomainDelete;
